Add release scripts (#32)
* Add release scripts

This commit adds support for releasing the operator. It adds a couple of scripts that are supposed to be executed by travis. The scripts build and push docker images, as well as change the 'deploy/operator.yaml' to use this latest release tag. Once that is done, the change to the operator.yaml is committed, a release is tagged and pushed to GitHub.

Signed-off-by: Juraci Paixão Kröhling <juraci@kroehling.de>

* Fixed image and github namespaces and renamed commit to version

Signed-off-by: Juraci Paixão Kröhling <juraci@kroehling.de>