cassandra-schema-job: set deadline to 1 day
If the job does not succeed within that duration
it transitions into a permanent error state.

That change is expected to smoothen deployment in
in environments where Cassandra takes a little less
predictable amount of time until it is available.

So far, in those environments the dealine of just
two minutes hits in frequently and then a human needs
to re-schedule the job to address a the underlying
usually *transient* problem which is more likely
to heal itself automatically with this patch.

Also see
https://github.com/jaegertracing/jaeger-kubernetes/pull/125
https://github.com/jaegertracing/jaeger-kubernetes/issues/32

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>