Completely replace the sidecar on each reconciliation, call patch instead of update. (#1212)
This is an attempt to stabilize the injection mechanism, before this patch we created a new sidecar and compare the new sidecar with the old one. The issue with this approach is the function that creates the sidecar uses the length of `dep.Spec.Template.Spec.Containers` to derive part of the arguments (See https://github.com/jaegertracing/jaeger-operator/blob/master/pkg/inject/sidecar.go#L198) . So in the first reconciliation process it inject one sidecar, but the next reconciliation process will inject a different sidecar with an extra argument. This triggers the pod re-creation process twice.

This approach recreated the sidecar in each deployment reconciliation

This also uses patch instead of update in an effort to alleviate some errors related to the version of the resource.

Fixes #1089

Signed-off-by: Ruben Vargas <ruben.vp8510@gmail.com>
Signed-off-by: Prageeth Warnak <prageeth.warnak@team.telstra.com>