Adjusted OLM manifests for 1.20, adjusted release script (#1226)
After the release of 1.20, prior to updating the Operator Hub, I noticed that the operator-sdk didn't automatically bump some fields for 1.20.

As a result, I added a few extra sed commands to bump the versions to the release script and ran the commands separately against the existing manifests, making them up to date for 1.20.

﻿Signed-off-by: Juraci Paixão Kröhling <juraci@kroehling.de>