Add GRPC Plugin support
Add support for GRPC plugins using an init-container image which
installs a plugin into an emptyDir volume on startup.

Example CRD for deployment with allInOne and a pre-existing secret
`plugin-config`:

jaegertracing.yaml

    apiVersion: jaegertracing.io/v1
    kind: Jaeger
    metadata:
      name: with-grpc-plugin
    spec:
      storage:
	type: grpc-plugin
	grpcPlugin:
	  image: jaegertracing/jaeger-operator-demo-storage-plugin:1.18.1
	  binary: "/plugin/demo-storage-plugin"
	  configurationFile: "/etc/plugin-config/config.json"
      volumeMounts:
      - name: plugin-config-volume
	mountPath: /etc/plugin-config
      volumes:
      - name: plugin-config-volume
	secret:
	  secretName: plugin-config

The specified container image will run as an init container and must
install the plugin binary to the location defined as `binary` above. The
binary must be installed in the directory /plugin.

Example Containerfile for installing the image:

    FROM registry.access.redhat.com/ubi8-minimal
    COPY --from=build /src/jaeger-storage /
    CMD ["cp", "/jaeger-storage", "/plugin/humio-jaeger-storage"]

Fixes #1042

Signed-off-by: Carl Henrik Lunde <carl.henrik.lunde@sparebank1.no>