Simplify OTEL related environment variables (#1255)
Signed-off-by: Kevin Earls <kearls@redhat.com>

This PR reduces the number of OTEL related EVs used by tests to two, and separates their functions.  If SPECIFY_OTEL_IMAGES is true the tests will explicitly add otel image names to CRs.  If SPECIFY_OTEL_CONFIG is true we will remove config values that are not used by the OTEL collector or ingester.